Split.split/2 duplicates leading mentions across splits
When posting a longer update in response to an update on Twitter or
Mastodon, the mention at the front of the update wasn't duplicated
across the splits.

On Twitter, this caused the first tweet to mention the author of the
tweet it's replying to, but the rest of the updates were replies to each
other, and dropped the mention.

This patch duplicates any mentions from the start of the update across
all splits, so each one mentions the user the first one is replying to.